dtoConverter and Candidate Service Working

+++ test/src/main/java/com/testbackend/test/models/entities/CandidateByTechnology.java
@@ -22,5 +22,5 @@ public class CandidateByTechnology {
     private Long idCandidateByTechnology;
     private Candidate candidate;
     private Technology technology;
-    private Integer experience;
+    private Long experience;
 }
